@@ -3,14 +3,14 @@ description: Demonstrates how to use the flutter_base plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: 1.0.2
-  cached_network_image: 2.5.1
+  cupertino_icons: 1.0.3
+  cached_network_image: 3.0.0
 
   flutter_base:
     path: ../
