@@ -5,28 +5,28 @@ author: archko
 homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: 1.0.2
-  json_annotation: 4.0.0
+  cupertino_icons: 1.0.3
+  json_annotation: 4.0.1
 
-  dio: 3.0.10
-  pull_to_refresh: ^1.6.4
+  dio: 4.0.0
+  pull_to_refresh: 2.0.0
 
   isolate: ^2.0.3
 
   provider: 5.0.0
   #webview_flutter: ^0.3.19+5
-  flutter_webview_plugin: ^0.3.11
+  flutter_webview_plugin: 0.4.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  json_serializable: 4.0.2
+  json_serializable: 4.1.3
 
 # The following section is specific to Flutter.
 flutter:
